% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_theta.R
\name{calc_theta}
\alias{calc_theta}
\title{Calculates theta, the angle of rotation, for a particular cross section segment.}
\usage{
calc_theta(xs)
}
\arguments{
\item{xs}{A select_xs (see ?select_xs) or list_xs (see ?list_xs) object.}
}
\value{
A data table with a column entitled "theta"
}
\description{
Calculates the angle of rotation, theta, for a cross section segment.
}
\examples{

}
\keyword{theta}
