% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_chem.R
\name{label_chem}
\alias{label_chem}
\title{Label the cross section with chemistry data}
\usage{
label_chem(distance, gs_elev, depth, tos, bos, chem_id, chem)
}
\arguments{
\item{distance}{Horizontal distance of each of the boreholes and/or wells along the cross
section. Should be a vector of depths. Defaults to NULL.}

\item{gs_elev}{The ground surface elevation at each of the boreholes and/or wells along the
cross section. Should be a vector of depths. Defaults to NULL.}

\item{depth}{The total depth (below ground surface) at each of the boreholes and/or wells
along the cross section. Should be a vector of depths. Defaults to NULL.}

\item{tos}{The distance from the ground surface to the top of the screened interval for
each well along the cross section. Should be a vector of depths. Defaults to NULL.}

\item{bos}{The distance from the ground surface to the tbottom of the screened interval for
each well along the cross section. Should be a vector of depths. Defaults to NULL.}

\item{chem_id}{The names of the desired chemical labels. The names specified
in chem_id should correspond with the column names of chem (see 'chem' below).}

\item{chem}{A data.frame of the chemical data to be plotted. The data.frame
should contain columns corresponding to the chemical parameters that will
be used to label the data.}
}
\value{

}
\description{
This function labels each sampling location shown on the cross section with
analytical (or other) data.
}
\examples{

}
\keyword{chemical}
\keyword{label,}
