% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_xs.R
\name{list_xs}
\alias{list_xs}
\title{Create a list of user specified cross sections}
\usage{
list_xs(xs_list, plot = FALSE, basemap = NULL, col = "blue")
}
\arguments{
\item{xs_list}{A list of cross sections objects developed using select_xs
(see ?select_xs)}

\item{plot}{Should a the plan view basemap be plotted? Defaults to FALSE.}

\item{basemap}{A data.frame or data.table containing x and y coordinates that
can be used to draw the basemap (see ?plot_bm)}

\item{col}{Color(s) that the line used to indicate the specified cross section
should be plotted. Can be a vector or an atomic value. Defaults to "blue".}
}
\value{
A data.table of the cross sections and their start, pivot, and end
  points.
}
\description{
This creates takes a list of cross section objects developed using select_xs
and binds them together into a data.table. Will plot the cross sections
on a plan view basemap if desired.
}
\examples{

}
\keyword{}
