% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bm.R
\name{plot_bm}
\alias{plot_bm}
\title{Plot a plan view of the basemap}
\usage{
plot_bm(basemap)
}
\arguments{
\item{basemap}{A data.frame or data.table containing x and y coordinates that
can be used to draw the basemap.}
}
\value{

}
\description{
This function plots a plan view of the basemap
}
\examples{

}
\keyword{basemap}
\keyword{bm,}
