% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_xs.R
\name{plot_xs}
\alias{plot_xs}
\title{Set up the plotting region for the cross section}
\usage{
plot_xs(data = NULL, distance = NULL, depth = NULL, gs_elev = NULL,
  name = NULL, xs_params = xs_params())
}
\arguments{
\item{data}{A data.frame or data.table with columns of "distance" (the maximum desired
distance of the cross section), "depth" (the maximum depth below ground surface), and
"gs_elev" (ground surface elevation). This parameter is not necessary if the distance,
depth, and gs_elev parameters are individually specified (see below). Defaults to NULL.}

\item{distance}{Horizontal distance of the boreholes and/or wells along the cross section or
the maximum distance that the cross section should extend. Can be a single value or a
vector of distances. If a vector of values is supplied, the maximum value is selected.
Defaults to NULL.
Note, the distance parameter is rounded to the nearest h_major unit (see ?xs_params) and
the x-axis limits are determined using the 'pretty' function (see ?pretty) with an
interval of 0 to the rounded distance value.}

\item{depth}{The total depth of the boreholes and/or wells below ground surface, or the
the maximum distance below ground surface that the cross section should extend to. Can be
a single value or a vector of depths. If a vector of depths is supplied, the maximum value
is selected. Defaults to NULL.
Note, the depth parameter is rounded to the nearest v_major unit (see ?xs_params)
and the y-axis limits are determined using the 'pretty' function (see ?pretty) with
the upper limit being the ground surface elevation (see gs_elev) and the lower limit
being gs_elev - the rounded depth value.}

\item{gs_elev}{The ground surface elevation at each of the boreholes and/or wells along the
cross section. Can be a single value or a vector of depths. If a vector of ground
surface elevations is supplied, the maximum value is selected. Defaults to NULL.
Note, the gs_elev parameter is rounded to the nearest v_minor unit (see ?xs_params)
and the y-axis limits are determined using the 'pretty' function (see ?pretty) with
the upper limit being the rounded gs_elev value and the lower limit
being the rounded gs_elev - and the depth (see depth).}

\item{name}{The cross section ID (e.g., A, B, etc.). Defaults to NULL.}

\item{xs_params}{A list of cross section parameters from the xs_params function (see
  ?xs_params). Defaults to xs_params().}
}
\value{
A plot of an empty gridded cross section.
}
\description{
This function sets up the plotting region for the cross section. Inputs are the total
length of the cross section, the depth below ground surface of the boreholes and/or wells
(or the maximum desired depth below ground surface that the cross section should extend),
and the ground surface elevation of the boreholes and/or wells. Using the supplied
information, and the cross section parameters set up using the xs_params function, an empty
gridded cross section will be plotted.
}
\examples{
plot_xs(distance = c(100, 200, 600), depth = c(30, 15, 55), gs_elev = c(10, 12, 11),
name = "A", xs_params = xs_params())
}
\keyword{cross}
\keyword{depth,}
\keyword{distance,}
\keyword{gs_elev,}
\keyword{plot,}
\keyword{section,}
\keyword{xs,}
\keyword{xs_params}

