% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_xs_locs.R
\name{plot_xs_locs}
\alias{plot_xs_locs}
\title{Plot wells/boreholes on the cross section}
\usage{
plot_xs_locs(dt = NULL, location = NULL, type = NULL, distance = NULL,
  gs_elev = NULL, depth = NULL, tos = NULL, bos = NULL,
  plot_gs = TRUE, w_width = 0.1, label_shift = 2, xs_par = xs_params())
}
\arguments{
\item{dt}{A data.frame or data.table with columns of "location", "type", "distance",
"gs_elev", "depth", "tos", and "bos". For explanations of these parameters and their
valid values, see the definitions below. The data parameter is not necessary if the
location, type, distance, gs_elev, depth, tos, and bos parameters are individually
specified. Defaults to NULL.}

\item{location}{A unique identifier for each borehole and/or well (e.g., "MW-3", "BH-45",
etc.). Should be a vector of character strings. Defaults to NULL.}

\item{type}{Is the specified location a well, borehole, etc.? Should be a vector of character
strings such as "well", "borehole", "ssp", "svp", etc. Defaults to NULL.}

\item{distance}{Horizontal distance of each of the boreholes and/or wells along the cross
section. Should be a vector of depths. Defaults to NULL.}

\item{gs_elev}{The ground surface elevation at each of the boreholes and/or wells along the
cross section. Should be a vector of depths. Defaults to NULL.}

\item{depth}{The total depth (below ground surface) at each of the boreholes and/or wells
along the cross section. Should be a vector of depths. Defaults to NULL.}

\item{tos}{The distance from the ground surface to the top of the screened interval for
each well along the cross section. Should be a vector of depths. Defaults to NULL.}

\item{bos}{The distance from the ground surface to the tbottom of the screened interval for
each well along the cross section. Should be a vector of depths. Defaults to NULL.}

\item{w_width}{Width of the well screen in inches. Defaults to 0.1.}

\item{label_shift}{The vertical distance by which to shift the borehole and/or well
labels. Defaults to 2.}

\item{plot_gs.}{Should a line be plotted between the ground surface elevations (gs_elev)
of the boreholes and wells? Defaults to TRUE.}

\item{xs_params}{A list of cross section parameters from the xs_params function (see
  ?xs_params). Defaults to xs_params().}
}
\value{
Plots the wells and boreholes on the cross section set up using grid_xs().
}
\description{
This function plots wells and/or boreholes on the cross section. The function will plot a
rectangular screened interval for wells. The bottom of a well is denoted with a dash and
the bottom of a borehole is denoted with a point.
}
\examples{
grid_xs(distance = c(100, 200, 550),
        depth = c(30, 15, 55),
        gs_elev = c(10, 12, 11),
        name = "A", xs_par = xs_params())
plot_xs_locs(location = c("MW-1", "BH-5", "CPT-3"),
             type = c("well", "borehole", "cpt"),
             distance = c(100, 200, 550),
             gs_elev = c(11, 8, 12),
             depth = c(60, 42, 90),
             tos = c(45, NA, NA),
             bos = c(55, NA, NA))
}
\keyword{cross}
\keyword{plot,}
\keyword{section}
\keyword{xs,}

