% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xs_plan.R
\name{plot_xs_plan}
\alias{plot_xs_plan}
\title{Plot cross sections on a plan view basemap}
\usage{
plot_xs_plan(xs, col = "blue")
}
\arguments{
\item{xs}{A select_xs (see ?select_xs) or list_xs (see ?list_xs) object.}

\item{col}{Color(s) that the line used to indicate the specified cross section
should be plotted. Can be a vector or an atomic value. Defaults to "blue".}
}
\value{

}
\description{
Plots a select_xs object or a list_xs object on a plan view basemap. This
function only plots the cross section line segments. Assumes a plan view
basemap has already been plotted.
}
\examples{

}
\keyword{xs}
