% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{2D rotation about an arbitrary point in space}
\usage{
rotate(coords, origin = c(0, 0), theta, units = "rad")
}
\arguments{
\item{coords}{A data.frame or data.table of x and y values (in feet; this
function will not work for latitude and longitude values that are in degrees).}

\item{origin}{The x and y value around which coords will be rotated. Defaults
to c(0, 0).}

\item{theta}{The angle of rotation through which the x and y values should be
rotated. The angle should be in radians or degrees.}

\item{units}{The units for theta. Should be radians ("rad") or degrees ("deg").
Defaults to "rad".}
}
\value{
A data.frame with the rotated x and y coordinates
}
\description{
Rotates a x and y coordinates by theta (the angle of rotation) about an
arbitrary point in space. Rotates the coordinates clockwise towards the
horizontal positive x-axis.
}
\examples{
df <- data.frame("x" = 1:5, "y" = 1:5)
rotate(df, theta = pi / 4)
rotate(df, origin = c(5, 5), theta = 90, units = "deg")
}
\keyword{rotate,}
\keyword{theta}
