% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_xs.R
\name{select_xs}
\alias{select_xs}
\title{Use a mouse to specify the desired cross section}
\usage{
select_xs(name, plot = FALSE, basemap = NULL, col = "blue")
}
\arguments{
\item{name}{The name of the cross section.}

\item{plot}{Should a the plan view basemap be plotted? Defaults to FALSE.}

\item{basemap}{A data.frame or data.table containing x and y coordinates that
can be used to draw the basemap (see ?plot_bm)}

\item{col}{Color that the line used to indicate the specified cross section
should be plotted. Defaults to "blue".}
}
\value{
A data.table with the name and the x and y coordinates of the specified
  cross section.
}
\description{
This function allows the user to specify the start, pivot, and end points
of the cross section by clicking the mouse pointer on locations shown on
a plan view basemap of the Site. After selecting the desired points, the
function will plot the selected cross section on a plan view basemap of the
Site.
This function returns a data.table with the name and the x and y coordinates
of the specified cross section.
}
\examples{

}
\keyword{xs}
