% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{stopServer}
\alias{stopServer}
\title{stopServer}
\usage{
stopServer()
}
\description{
Stops the server used by blogdowns server_site
}
\examples{
# Use the addins browser or activate via the addins menue to
# stop the server

}
