% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/xs_params.R
\name{xs_params}
\alias{xs_params}
\title{Specify the cross section parameters}
\usage{
xs_params(v_scale = 10, h_scale = 100, v_minor = 1, v_major = 10,
  h_minor = 20, h_major = 100, units = "ft", plot_v_minor = TRUE,
  plot_v_major = TRUE, plot_h_minor = FALSE, plot_h_major = TRUE)
}
\arguments{
\item{v_scale}{Vertical scale of the y-axis. Defaults to 10 (i.e.,
in the resulting plot, 10 feet on the y-axis will correspond to 1 inch printed).}

\item{h_scale}{Horizontal scale of the x-axis. Defaults to 100 (i.e.,
in the resulting plot, 100 feet on the x-axis will correspond to 1 inch printed).}

\item{v_minor}{Interval for plotting minor horizontal grid lines. Defaults to 1 (i.e.,
a minor gridline every foot).}

\item{v_major}{Interval for plotting major horizontal grid lines. Defaults to 10 (i.e.,
a major gridline every 10 feet).}

\item{h_minor}{Interval for plotting minor vertical grid lines. Defaults to 20 (i.e.,
a minor gridline every 20 feet).}

\item{h_major}{Interval for plotting major vertical grid lines. Defaults to 100 (i.e.,
a major gridline every 100 feet).}

\item{units}{Unit of measurement for the x and y axes. Defaults to feet ("ft").}

\item{plot_v_minor}{Plot minor horizontal grid lines? Defaults to TRUE.}

\item{plot_v_major}{Plot major horizontal grid lines? Defaults to TRUE.}

\item{plot_h_minor}{Plot minor vertical grid lines? Defaults to FALSE.}

\item{plot_h_major}{Plot major vertical grid lines? Defaults to TRUE.}

\item{h_major}{Interval for plotting major vertical grid lines. Defaults to 100 (i.e.,
a major gridline every 100 feet).}
}
\value{
List of parameters that will be used to set desired cross section parameters.
}
\description{
This function allows you to specify desired cross section parameters such as
horizontal and vertical scales and gridlines.
}
\examples{
xs_params()
}
\keyword{cross}
\keyword{params,}
\keyword{section}
\keyword{xs,}

