% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_xs_locs.R
\name{plot_xs_locs}
\alias{plot_xs_locs}
\title{Plot wells/boreholes on the cross section}
\usage{
plot_xs_locs(data = null, location = null, type = null, distance = null,
  gs_elev = null, depth = null, tos = null, bos = null,
  plot_gs = TRUE, w_width = 0.1, label_shift = 10,
  cross_section_params = cross_section_params())
}
\arguments{
\item{v_scale}{Vertical scale of the y-axis. Defaults to 10 (i.e.,
in the resulting plot, 10 feet on the y-axis will correspond to 1 inch printed).}

\item{h_scale}{Horizontal scale of the x-axis. Defaults to 100 (i.e.,
in the resulting plot, 100 feet on the x-axis will correspond to 1 inch printed).}

\item{v_minor}{Interval for plotting minor horizontal grid lines. Defaults to 1 (i.e.,
a minor gridline every foot).}

\item{v_major}{Interval for plotting major horizontal grid lines. Defaults to 10 (i.e.,
a major gridline every 10 feet).}

\item{h_minor}{Interval for plotting minor vertical grid lines. Defaults to 20 (i.e.,
a minor gridline every 20 feet).}

\item{h_major}{Interval for plotting major vertical grid lines. Defaults to 100 (i.e.,
a major gridline every 100 feet).}

\item{h_major}{Interval for plotting major vertical grid lines. Defaults to 100 (i.e.,
a major gridline every 100 feet).}

\item{units}{Unit of measurement for the x and y axes. Defaults to feet ("ft").}

\item{plot_v_minor}{Plot minor horizontal grid lines? Defaults to TRUE.}

\item{plot_v_major}{Plot major horizontal grid lines? Defaults to TRUE.}

\item{plot_h_minor}{Plot minor vertical grid lines? Defaults to FALSE.}

\item{plot_h_major}{Plot major vertical grid lines? Defaults to TRUE.}
}
\value{
List of parameters that will be used to set desired cross section parameters.
}
\description{
This function plots wells and/or boreholes on the cross section
}
\examples{
xs_params()
}
\keyword{cross}
\keyword{params,}
\keyword{section}
\keyword{xs,}

